[cc] Cache scrollbar location in UpdateThumbAndTrackGeometry
The geometry returned by scrollbar_ will vary dramatically depending
on previous function calls (in particular,
RenderLayer::updateLayerPositions and RenderLayer::paintOverflowContents
will set the values returned to very different things).

The scrollbar offset is computed using scrollbar_->TrackRect() and
scrollbar_->Location(), but TrackRect() uses a cached value, which ends
up being incompatible with Location().

Cache the result of scrollbar_->Location() so that a consistent set of
values is used. Also, it appears that it isn't valid to call 
scrollbar_->Location in PushPropertiesTo to begin with.

BUG=271914

Review URL: https://chromiumcodereview.appspot.com/22978033