App: Make sure unexpected exit paths generate crash dumps for any process.
Right now it is possible for some random code to directly terminate some
process (most notably the browser) without generating a crash dump.

This CL moves r217049 to a place where it covers any process type while
at the same type is Chrome-specific (not part of content). It is a 'minimal
patch' designed to be applied to and older branch.

BUG=275818
TEST=none
R=thestig@chromium.org

Review URL: https://chromiumcodereview.appspot.com/23190042