Remove the Extensions URLRequestContext.
Though chrome-extension: scheme URLs support cookies, they do not share
a namespace with http: and https:. In particular, chrome-extension://a and
http://a should not have the same set of cookies.

To enforce this, previously the code created a completely separate
URLRequestContext for servicing chrome-extension: schemes. However,
the code really only used this object as a method for conveying the
correct cookie jar from Profile creation to a few spots where cookies
were accessed; the rest of the URLRequestContext functionality was unused.

This CL removes the Extensions URLRequestContext code and replaces it
with APIs that directly expose the needed net::CookieStore.

Lastly, CookieMonster::EnableFileScheme() is removed and 
CookieMonster::Delegate is renamed CookieMonsterDelegate.

EnableFileScheme is an inherently racy API because 
CookieMonsters are creatable on all threads and this 
function sets an unprotected global flag. CookieMonsterDelegate
is preferable to the nested interface because it can now be
forward declared.

TBRing darin and sky to cover the rest of the mechanical unittest changes.

TBR=darin,sky
BUG=158386,159193,57884

Review URL: https://chromiumcodereview.appspot.com/12546016