Upstream TestHttpServerClient for Android.
The TestHttpServerClient is a test utility class to get URLs that point
to the test HTTP server, and to verify that it is up and giving expected
responses to given URLs.

This class used to live downstream, but it is required for upstreaming
the sync tests.

As part of this, the StreamUtil is also upstreamed.

BUG=272584

Review URL: https://chromiumcodereview.appspot.com/23437002