Skip clearing the canvas before painting the contents of opaque layers
The SkCanvas doesn't need to be cleared before painting the layer contents
if the layer is known to be opaque (e.g. the root layer). Bypassing the clear
gives a sizeable perf boost.

This patch also switches to calling clear() instead of drawRect() for clearing the canvas pixels as clear has been shown to be significantly faster. 

BUG=279612

Review URL: https://chromiumcodereview.appspot.com/23484005