[Files.app] Not to show the error dialog of unavailable folder shortcut.
This patch is a part of https://codereview.chromium.org/23420003/. The original patch is too large to merge to branch, so we created this small patch only to remove the error dialog of unavailable folder shortcut.

This is a temporary patch. Once we confirm the fix on the M30 branch, we'll commit the reaming part of the original patch.

BUG=270436
TEST=manual
TBR=hirono@chromium.org

Review URL: https://codereview.chromium.org/23669003