[sync] Don't modify text of advanced sync button unless contents have changed
The chrome settings page updates the text of the advanced sync settings
button every time it is displayed. This can affect the search-highlight
state of the button when an overlay is closed and the settings page is
re-displayed.

This patch causes the settings page to modify the button text only when
the contents have changed, so that search-highlight behavior is not
affected.

BUG=268265
TEST=Search for "settings" in the settings page and open/close the
advanced sync dialog. Highlighting must remain as it was.

Review URL: https://chromiumcodereview.appspot.com/23496007