cc: Cleanup offscreen context plumbing.
When I added offscreen context provider, I plumbed it to the renderers
through the ResourceProvider. This was a bad idea, and using the
ResourceProvider as a suitcase for plumbing stuff is poor practice.

Secondly, we have a funny round trip from LayerTreeHostImpl to
ThreadProxy::DidTryInitializeRendererOnImplThread which just
goes back through the LayerTreeHostImpl to set the offscreen
contexts.

This patch removes the offscreen ContextProvider from the
ResourceProvider and the removes the DidTryInitializeRendererOnImplThread
method from ThreadProxy. Instead we have SetOffscreenContextProvider()
on the LayerTreeHostImpl, and the LTHI owns the provider and gives
it directly to the renderer for each frame. The renderer holds onto the
provider only while drawing its frame.

While we're at it, make pixel tests faster (~20ms or 10% on my z620) by
only making an offscreen context for tests that want one.

Discussion around this CL also led me to the conclusion that
RendererClient is evil and we should pass things directly into
the Renderer rather than have it call out for them.

R=boliu, piman
BUG=

Review URL: https://chromiumcodereview.appspot.com/23660002