Deprecates ContinueGetResourceList and introduces GetRemaining{File,Change}List.
On Drive API v2, ResourceList in GData WAPI is split into two, FileList and
ChangeList. This CL is the preparation to use Drive API v2 classes.
Note that this is also the preparation to fetch feeds with only fields we're
interested in.
The new methods are not yet used from production. It'll be done in a separate
CLs.

BUG=277253
TEST=Ran unit_tests

Review URL: https://chromiumcodereview.appspot.com/23450009