Try to detect situations where the bisect script can still dive in to V8's "bleeding edge" repository and get back a meaningful result.
At the moment, it tries to detect when both the current revision and previous revisions on trunk are just normal pushes from bleeding edge, and not individual merges. It checks that the description is "Version X.Y.Z" for svn revision N, and then checks that there's a CL in bleeding edge with a description "Prepare to push to trunk. now working on X.Y.(Z+1)".

In any other case, the script won't try to dive in any further and just report the version on trunk that it narrowed down to.


BUG=279493
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/23620005