ErrorConsole Switch, Duplicate Error Detection, Dev Mode Check
Add in a command-line switch for ErrorConsole ('enable-error-console').
Add in duplicate-error detection, so that multiple copies of the same error aren't both stored.
Add in a check so that the ErrorConsole only stores errors if the user is in Developer Mode.
Rename ManifestParsingError to ManifestError, JavascriptRuntimeError to RuntimeError

BUG=21734

Review URL: https://chromiumcodereview.appspot.com/23311011