PNaCl test: Use "private" version of nacl_exception library to bypass IRT
Since I changed the PNaCl shims to filter out IRT interfaces that are
not supported under PNaCl (in
http://src.chromium.org/viewvc/chrome?view=revision&revision=213741),
this test no longer tests that the nacl_exception syscalls are
disabled under PNaCl.  It passes because the nacl_exception IRT
interface is blocked.

Changing the test to use the "private" version of the library means it
uses the syscalls directly, which ensures we test whether the
nacl_exception syscalls are disabled under PNaCl.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3574
TEST=browser_tests

Review URL: https://chromiumcodereview.appspot.com/23020028