[Android] Rename SurfaceTextureBridge to just SurfaceTexture.
This came up in https://codereview.chromium.org/22912020/ but was punted
to keep merging smaller. Now that it's in, here's the change. The java
class has a slightly different name as internally to the native
SurfaceTexture object we have a wrapper, but that detail is unknown to
clients.

Review URL: https://chromiumcodereview.appspot.com/23574003