Clear render surface when we early out in CalcDrawPropsInternal
If we don't clear the layer's render surface when we early out, we'll build the
render surface layer list incorrectly, and the layer iterator will crash trying
to iterate over it. I've done this at the first early out, but I missed this
at the second.

R=danakj@chromium.org
BUG=280613

Review URL: https://chromiumcodereview.appspot.com/23708003