Fixes the display notification's handling of login status.
The current code refers the login status for when the notification
is created, but it may be changed after created. Rather OpenSettings
should verify the current login status.

This is a potential issue but not revealing seriously right now,
since the notification popups are dismissed when the screen is locked,
but the current code may cause a concern in the future we will
keep some of system notifications.

Also this CL removes unused class. It should have been removed.

BUG=None
R=stevenjb@chromium.org

Review URL: https://chromiumcodereview.appspot.com/23710006