Allow speech recognition to work in content_shell.
Until now, content_shell was exposing the Web Speech API's recognition feature
to websites, but if they tried to use it, nothing happened because content_shell
doesn't know how to ask the user for permission to use the microphone. To fix
this, a SpeechRecognitionManagerDelegate was implemented for content_shell to
grant speech recognition permission when requested.

BUG=237119

Review URL: https://chromiumcodereview.appspot.com/23606003