Add a GYP command to GN. This runs GYP and then GN, and does some work to integrate the builds.
Adds an "external" variable that's set in the GYP mode of GN that will cause it to not write out a target but to pretend it's there. These are the ones that should be generated by GYP. This value is not used in the normal "gen" mode so we actually have two builds: a separate GN-only build that includes GN itself and some other low-level libraries, and a hybrid build.

This renames and moves some of the GN targets so that the names and locations match those of the GYP build.

I prefixed the ninja rules for the default toolchain to avoid colliding with the GYP ones. I also fixed some places where the rules were not getting prefixed.

I added an exception to target output labeling to avoid double-prefixing with "lib".

TBR=scottmg
BUG=

Review URL: https://codereview.chromium.org/23493013