Always SetAllowOverlappingViews for tab contents.
This has the effect of un-doing
https://chromiumcodereview.appspot.com/16140023/
because it turns out that the don't-allow-overlapping-views path
is more broken than allow-overlapping-views path.

In particular, the issue is that the GL context is removed from
the view immediately, and whatever content was there before (the
last non-accelerated frame) is revealed until a new non-accelerated
frame is drawn.

BUG=279472

Review URL: https://chromiumcodereview.appspot.com/23629006