Override WebPreferences::force_compositing_mode in layout tests.
In layout tests, WebPreferences::force_compositing_mode should be as set in the layout_tests configuration not as declared by chrome's compositor_util.

As suggested @ http://crbug.com/267038#c30 and in offline discussions.

This is overridden as a result of:
    content::ApplyLayoutTestDefaultPreferences()
    WebKitTestController::OverrideWebkitPrefs()
    ShellContentBrowserClient::OverrideWebkitPrefs()
    WebContentsImpl::GetWebkitPrefs()
which is called from WebContentsImpl::GetWebkitPrefs() after the typical preference is set and thus results in the desired override (also note that ShellContentBrowserClient::OverrideWebkitPrefs() is a no-op in non-layout-tests world as desired).

BUG=267038

Review URL: https://chromiumcodereview.appspot.com/23772002