[temporary M30 fix] Disable dragging tabs with web contents modal dialogs displayed
The full fix for the crashes in the associated bugs in trunk is too
risky to merge into M30. After discussions among kareng and sky, the
best approach was to disable dragging of tabs while a web contents modal
dialog is displayed, to prevent the crashes. This only affects Chrome
OS and will be reverted after testing in Canary and merging to the
branch.

BUG=275274, 274856
R=sky@chromium.org

Review URL: https://codereview.chromium.org/23698003