Added basic browser test for window.print() from Chrome apps.
This just tests the success of the call to window.print(). It does not
test whether the call leads to a print preview window and ultimately to
a printout. It will be extended over time to provide additional coverage
of the print functionality.

BUG=NONE

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=220467

Review URL: https://chromiumcodereview.appspot.com/23054002