Add quit item to app menu.
This also changes the action associated with Quit Chrome when an app
is focused. This is to allow Quit <app> to have the same key binding.
The action closes the app instead of attempting to quit Chrome.

BUG=168080, 276052
TEST=Focus an app's window.
The main menu bar should contain a menu for the app which contains a Quit item.
Selecting the Quit item, or using Cmd+Q quits the app.
TEST=Focus an app's window.
Right click Chrome's dock icon and select Quit.
Chrome and all apps should quit.

Review URL: https://chromiumcodereview.appspot.com/23301018