[Android] Support startup histogram.
This CL adds support for the
Startup.BrowserMessageLoopStartTimeFromMainEntry UMA histogram to
chromium_testshell. It does this by recording the start time as
early as possible on the Java side (well before the C++ library is even
loaded), and then reading this on the C++ side through a new JNI call.

BUG=266173

Review URL: https://chromiumcodereview.appspot.com/23443009