Docserver: SidenavDataSource refactor, transition to DataSourceRegistry
The next step in moving towards the DataSourceRegistry architecture
(see bug).

The SidenavDataSource is the first DataSource to rely on the contents of
a request and the DataSourceRegistry had to change. Now DataSources are
given both a ServerInstance and a Request when created. Strings and
Manifest DataSources have been updated.

SidenavDataSource contained some very old code and was in need of
refactoring. The Factory had to be stripped out because it is now
unneeded and doesn't work with the registry.

CreateDataSource was moved into the RenderServlet and is added to the
template data source in Create to give the DataSources access to 
a Request.

BUG=275039
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/22824042