Add UI for network configuration when offline during kiosk app launch.
This is a first pass at implementing UI features that would allow the user
to configure the network during a kiosk app launch. A small button appears on
the splash screen after a short period if network connectivity is not detected.
Clicking this button brings up a signin screen for the owner of the device, and
after authenticating, the standard network error dialog appears.

BUG=242589
TEST=manually on device

Review URL: https://chromiumcodereview.appspot.com/23503014