Add ability to retrieve the width and height from a SKP file.
In trace-viewer we want to allow users to load up .skp files in the
skia_debugger.html. In order to do that we need access to the width and
height of the skp in order to get the op timings.

To that end, this CL adds a getInfo method to chrome.skiaBenchmarking to return
the width and height.

BUG=

Review URL: https://chromiumcodereview.appspot.com/23572018