[rAc] Only show matching hint icons when entering a new credit card.
Previously, the UI showed a generic set of four credit card issuers, and
highlighted the matching one as the user types.  This had several issues,
including not being fully comprehensive and at the same time encouraging users
to enter AmEx cards, which are not fully supported.  So, instead, we now only
show at most a single image, which exactly matches the card that the user is
entering.

BUG=255107
TEST=only one icon shown, icon matches actual card type
R=estade@chromium.org

Review URL: https://chromiumcodereview.appspot.com/23543009