[Telemetry] Allow Telemetry to run on a non-rooted device.
The problem was that Get/SetProtectedFileContents doesn't work if the file is
not actually protected. So we must first check whether it is protected before
using them.

I've tested this with Chrome and Content Shell on both a rooted and non-rooted
device. Everything works.

The one wrinkle is that if you run against full Chrome on a non-rooted device,
you have to manually enable devtools, but Telemetry prompts you for that.

BUG=

Review URL: https://chromiumcodereview.appspot.com/23649003