Move the directory of CLD2
To put the file README.chromium, OWNERS and so on, we need to move the directory of the third-party library.

We cannot assure that it is safe to move from 'third_party/cld2' to 'third_party/cld2/src', so we adopted the new path 'third_party/cld_2'.  See: https://codereview.chromium.org/23301019/

Note: the newer revision (r74) of CLD2 is specified because the newer version includes some security fixes.

BUG=240647

Review URL: https://chromiumcodereview.appspot.com/23449011