Ignore touch events when the touch press is on the side bezel region
We currently have a #enable-touch-side-bezels flag default to disable
any touches from bezel regions. The implementation is that if the flag
is not set (bezel not enabled), all X touch events at bezel region
are not propagated into aura.

One edge case is that when touch starts at non-bezel region, but ends
at bezel region, dropping the touch end events will make the tracking
in aura think the touch event is never released.

This patch changes that ignore touch events (bezel and non-bezel) if
the touch press is on the bezel region. So
1) If touch starts at bezel, then no matter the touch stays at side
bezel or moves to non-bezel, all the touches are ignored.
2) If touch starts at non-bezel, then moves into bezel/or releases
at bezel, the move & release are always NOT ignored.

BUG=277029
TEST=1. Use page http://www.rbyers.net/eventTest.html  for test.
     2. Start touch at non-side-bezel region, and move to the edge
        bezel and release. Make sure touch end is received. Start
	touch at side bezel region, and move the touch to non-bezel
        region, make sure touch move & release are not received.
     3. Repeat taping close to left bezel and right bezel at the same
        time. Make sure after many repeats, there is no stuck fingers.
        Follow crbug.com/274841 for instructions.

Review URL: https://chromiumcodereview.appspot.com/23295023