[Android] SelectActionBar - resolve style for share/web search icons.
Right now we hard code holo-light drawables for the share and web search
icons on the action bar, we should resolve the style as we do for the
other icons. Only snag is that the attributes are not currently public
in the Android framework, so we need an extra level of indirection.

BUG=283866

Review URL: https://chromiumcodereview.appspot.com/23503020