Enable GPU blacklist in tests.
Partial re-land of https://codereview.chromium.org/22198004/) -- Always enable FCM on Windows -- Part 1/3.

On top of fixing some of the issues there; only lands the part that enables the GPU blacklist in tests as the former CL is too hard to land all at once.

Also keeping --skip-gpu-data-loading around for now to be able to deal with failures caused in layout_tests by this on their own later (this needs to land ASAP and layout_tests don't need this ASAP).

Another CL will follow to always enable FCM on non-blacklisted Windows machines.

This part re-enables loading the blacklist in tests (and adds a content_browsertest to make sure that the configuration we expect to be testing is indeed the one we are testing -- this uncertainty is basically the only reason the blacklist was explicitly disabled before).

This CL also cleans up compositor_util.cc which was enforcing the blacklist twice.

The original plan was to do this only for Windows as Mac/Linux was causing trouble, but it turns out to be harder to do it only on Windows; so taking care of http://crbug.com/277242 in this CL too after all... 

BUG=233830, 267038, 190942, 277242
TBR=jcivelli, piman 

Originally Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=219132 
Then Reverted: https://src.chromium.org/viewvc/chrome?view=rev&revision=219159

Review URL: https://chromiumcodereview.appspot.com/23534006