sync: Remove IDs from OrderedCommitSet
Remove the IDs from OrderedCommitSet in order to add a new member,
AddCommitItems().

Removing the IDs from this class would have been a good idea on its own.
There hasn't been any need to track both IDs and metahandles for quite
some time, and storing redundant information always carries the risk
that the two copies will disagree with each other. It is particularly
risky in this case because the entries' IDs will be changed when a
commit response is successful.

The AddCommitItems() function would have needed to have an inconvenient
signature if it had to pass in both the handles and IDs of all the
entries being added to the set.  It's much easier to implement now that
we only need to pass in the entries' metahandles.

The new member function is not used in this commit.  It will be used in
a future commit that allows us to run GetCommitIdsCommand on a single
model type at a time.

BUG=278484

Review URL: https://chromiumcodereview.appspot.com/23694004