[Android] Provide a valid initial size for the overscroll effect
Chrome on Android's overscroll effect is created on-demand, the first time a
user overscrolls.  Rather than wait for an update from the compositor to give
the effect a valid size, provide the created effect the last valid content size.
This prevents the initial overscroll hitch that may occur if the compositor is
otherwise inactive when the first overscroll update is received.

BUG=284523
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/23494021