[Files.app] Refile a timing of doing selectBestMatchItem().
Previously, selectBestMatchItem() is called in the 'permuted'/'changed' handlers of the data model. But in this timing, although we can make sure that the data model is updated, but the list elements in the list may not be updated (it depends on the order of registration of event listeners). Then we can't select the target item since the target item may not be created yet.

This patch delays it just after the updating of elements in list.

BUG=280155
TEST=manual

Review URL: https://chromiumcodereview.appspot.com/23618028