Update googlestorage version of archive_layout_test_results.py
This patch changes the script to upload the layout-test-archives files
to googlestorage in two different ways. First, we upload the data to
two different locations: buildername/buildnumber and buildername/latest
so that the flakiness dashboard and garden-o-matic can have a stable url
for the "latest" results. Second, we upload the individual files as well
as the zip file, so we can link directly to an individual result as well
as download them all at once.

Eventually we should remove the "latest" version and update the tools
to always request specific versions, but for now we need this for compatibility.
If storage space ever becomes a concern we can revisit this.

R=hinoka@chromium.org, ilevy@chromium.org, xusydoc@chromium.org, iannucci@chromium.org
BUG=261749

Review URL: https://chromiumcodereview.appspot.com/23881003