Makes sure that an input method doesn't reference to a destroyed object.
This CL makes sure that an input method never reference to a destroyed object.  We're not 100% sure why Issue 280219 is happening, but it seems a destroyed object is referenced from an input method.

For render_widget_host_view_win.cc, there is no need to check the focus.  Even if it's not focused, the object going to be destroyed should be removed from the TSFBridge.

BUG=280219
TEST=none

Review URL: https://chromiumcodereview.appspot.com/23478020