On user-level Chrome self-destruct, make user-created shortcuts target system-level chrome.exe.
The old behavior was to delete all the shortcuts, which is rather user-hostile. With this change, the user-created shortcuts will be made to target the new system-level Chrome, to form a better transition experience.

- To determine whether or not a shortcut (to the user-level chrome.exe) is user-created, the heuristic we use is to test that the arguments of the shortcut are empty.
- The default shortcut to user-level chrome.exe is NOT retargeted, so it would get deleted (thus avoiding duplicate shortcut to the system-level Chrome).

Right now the "Chrome Apps" is not handled. This is a similar issue as http://crbug.com/238895 , and should be handled in a different CL.

BUG=235857

Review URL: https://chromiumcodereview.appspot.com/22382007