Fixes focus traversal bug
When on the last focusable view in a Container FocusManager would
start searching from Widget. Problem is child Widgets share the same
FocusManager, meaning focus in a child widget wraps to the parent,
which is unexpected in most situations. In theory there could be
situations where we should wrap to the parent, but I don't think we
have any.

BUG=276213
TEST=covered by test
R=dmazzoni@chromium.org

Review URL: https://chromiumcodereview.appspot.com/23475012