Fixes use after free during drag and drop.
Apparently under windows requesting focus can trigger capture lost,
which caused all sorts of problems with tab dragging. When you've
dragged enough to initiate a real tab drag we save focus. The code
never expected saving focus to trigger a capture lost (which stops tab
dragging), which triggered the crash. The fix is to make the code
allow for saving focus to delete the TabDragController.

I'm also converting from storing a raw View* for the previously focused view to a ViewStorage id. This way if the previously focused view is deleted no crashes.

I had to add a bit of infrastructure for the test.

BUG=275931
TEST=covered by test now.
R=ben@chromium.org

Review URL: https://chromiumcodereview.appspot.com/23523018