Show ****** in wifi passphrase field
Shill does not send the passphrase to Chrome for security reasons, so
show ****** in the passphrase field when configuring wifi if the service
is Connectable. (Sometimes when a connection fails we do not know why
it failed and so we show the configure dialog in case it is an error in
the configuration).

BUG=264819

Review URL: https://chromiumcodereview.appspot.com/23522013