Simple Cache: restore the high-res timer
This *reverts* the parts of high-res timing that was moved out from simple cache
in attempt to have it in base/, here:
http://src.chromium.org/viewvc/chrome?view=revision&revision=216478

The changes in base/ are untouched today. This partial revert should be safe
because this change just adds (un-removes) a different implementation and
switches simple cache to using it.

The bug being blocked by this deficiency is in the BUG field, we do not have a
specific bug for the problem, it would have required age restriction on the
audience.

BUG=261618

Review URL: https://chromiumcodereview.appspot.com/23819025