Mac: Fix window raising for multiple desktops
When raising a set of windows in response to a dock icon click, only raise the
ones on the currently active space. We currently raise all of them and switch
spaces haphazardly.

Also tidy up the applicationShouldHandleReopen:hasVisibileWindows: callback. We
should return NO since we've reacted and don't need AppKit to deminiaturize for
us. (Although it seems to not make much difference, probably because we've
already picked a window to deminiaturize, so AppKit won't pick a second.) Also
remove the tabbed/popup check. As of r192264, those are the only browser window
types.

BUG=281674
TEST=Open a browser window in Desktop 1 and one in Desktop 2.
Switch to Desktop 2.
Open another application in Desktop 2 and make it foreground.
Click the dock icon for Chrome. Should remain on Desktop 2.

Review URL: https://chromiumcodereview.appspot.com/23737003