Handle cache-control:no-store header in PNaCl translation cache
Pexe files with the cache-control:no-store header should not be cached.
Add a field to the PnaclCacheInfo struct, plumb the value all the way
from the plugin to the browser, and treat it basically the same way we
currently treat incognito translations (since we currently don't have an
off-the-record cache for those).
R=jvoung@chromium.org
BUG=none, noticed this was missing when doing cleanup

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=221275
Reverted due to iterator use-after-erase in patchset 4. fixed in 5. No idea why it passed the CQ so cleanly but blew up the buildbots so completely.

Review URL: https://chromiumcodereview.appspot.com/23458015