Allow NOLINT to skip the std::abs PRESUBMIT check.
shader code is one example where abs() without std:: is acceptable.
But we still would like to generally keep the PRESUBMIT check that
enforces all uses of abs() and related functions must be prefixed
with "std::"

So in order to prevent PRESUBMIT annoyances when we modify shader
code or other similar situations, the regular expression for the
presubmit is modified to recognize a "// NOLINT" comment. In that
case, it ignores any un-prefixed abs() usage.

BUG=261900

Review URL: https://chromiumcodereview.appspot.com/23591027