Set hbrBackground on wndclass, fixes white flash on resize
When in DWM, the DefWindowProc handling of WM_NCPAINT was clearing (to
white) in the area that was not yet drawn. When the wndclass's
background is set to COLOR_WINDOW this is avoided and the frame
appears to stay more "solid".

This only affects the white flash in the NC area in the linked bug,
not the grey in the content area. We should also consider reverting
the compositor.cc change from here http://crrev.com/192657 as it seems
to make resize a bit laggier.

R=sky@chromium.org
BUG=177115

Review URL: https://chromiumcodereview.appspot.com/23596011