This defers starting background extension page RenderViews until after session restore has completed.
This is mainly to help login/startup times so that Chrome is useful to
the user earlier.  To make sure that this is actually helping, this CL
includes a Finch experiment that will only enable the deferral on 50%
of the clients.

It is also expected that deferring these will help with some problems
we've seen when extensions attempt to do GAIA authentication.

BUG=279427,259791
TEST=Ran performance tests on ChromeOS and Linux

Review URL: https://chromiumcodereview.appspot.com/23618014