cc: Speed up CleanUpImageDecodeTasks a bit.
This patch keeps the state of used layers and updates it in register
and unregister tile. This eliminates the need for CleanUpImageDecodeTasks.

This also provides a fix to the memory limits in the perftest, so by
comparison the performance should drop significantly. However, this is
only because the perftests are now properly testing the full
ManageTiles code path. The patch to remove CleanUpImageDecodeTasks
actually improves performance by about 15%.

BUG=283777
R=reveman@chromium.org,tomhudson@chromium.org

Review URL: https://chromiumcodereview.appspot.com/23757024