Switch to doing user-resizing via system for panels on Mac
We used to track and handle user-resizing by ourselves. This brings up a few problems and makes the user-resizing hard to use on Mac. For example, the user-resizing could not be triggered outside or on the edge. Switching to doing it via system solve all these problems.

Also removed EnableResizeByMouse which is not really used in our code.

BUG=282652,282657
TEST=manual test by user-resizing panels

Review URL: https://chromiumcodereview.appspot.com/23918002