Add Hide to app menu.
This adds "Hide <app>" to the app menu that is shown when an app window
has focus. This also enables the Cmd+H shortcut. 

This also adds an IPC message to instruct the shim to hide.

This also adds a command ID for Hide and a corresponding tag to
MainMenu.xib.

BUG=168080, 276052
TEST=Start an app.
Press Cmd+H or select Hide from the main menu. The app should hide.
Clicking the app's dock icon, or Cmd+Tab to the app should show it again.

Review URL: https://chromiumcodereview.appspot.com/23514021