Initial Geolocation Wifi Scan Checking Fix for CrOS
The ChromeOS Wifi provider was erronously reporting an initial
scan of 0 wifi access points when the shill's scanning had not
yet completed. This resulted in a fallback to IP geo for the
first request and then proper Wifi based location on the
second request.

BUG=236418

Review URL: https://chromiumcodereview.appspot.com/23889009