Added discrepancy and inverse RMS frame time to smoothness benchmark.
Discrepancy is a measure for the largest area of jank in a series of time stamps.

Inverse RMS frame time is a generalization of frames/second that penalizes longer frame times. We can switch to the inverse of another generalized mean with a higher exponent if we want to penalize long frame times even more.

R=nduca@chromium.org, tonyg@chromium.org, klobag@chromium.org, vangelis@chromium.org
BUG=280627, 281544

Review URL: https://chromiumcodereview.appspot.com/23506030