CC: Fix missing swap-used-incomplete-tile updates
When we early-out of a frame that has no damage, we didn't
correctly detect if the frame had incomplete tiles. This
patch only updates the flag if a swap actually occurs.

To prevent this kind of bug in the future, this patch
always sets the did-swap-use-incomplete-tile flag when it is
known, rather than resetting it to false in one place and
hoping it will always be set again correctly in all cases.

The test times-out without this patch.

BUG=284810

Review URL: https://chromiumcodereview.appspot.com/23686011