[NaCl SDK] Don't exit from nacl_io_test before posting message to JS.
In the nacl_io_test there's a race condition between the NaCl module exiting
and a posted message from the module being handled by JavaScript.

To fix it, when running in the browser, we wait indefinitely and assume the
JavaScript will kill the module when it is ready.

(This fixes a flake that can be seen in
http://build.chromium.org/p/client.nacl.sdk/builders/mac-sdk-multi/builds/5838/steps/Run%20Tests/logs/stdio

BUG=none
R=sbc@chromium.org

Review URL: https://codereview.chromium.org/23531034