Fixes crah in NativeWidgetAura::IsActive()
The crash happens if when the HWND is destroyed someone asks if a
child NativeWidgetAura is Active. In particular during this code path
we've gone through ~RootWindow and are in ~Window. When this happens
GetRootWindow() uses Window's implementation, which returns NULL and
we crash.

BUG=264855
TEST=none
R=ben@chromium.org

Review URL: https://chromiumcodereview.appspot.com/23904005