Added support of relative mouse motion in Chromoting.
This CL makes the client plugin treat a completely transparent mouse cursor as an indication that the host switched to relative mouse mode.

If the browser rejects or cancels the mouse lock for any reason, the plugin sets the cursor to the standard arrow pointer until a new cursor is set by the host.

The webapp has to send the 'allowMouseLock' message to enable this behavior.

BUG=132613

Review URL: https://chromiumcodereview.appspot.com/23484015