[Activity log] Rework extraction of URLs from argument lists
Redo the code for handling the tabs API and tab ID translation to make
it more generic.  Make it table driven: there is now a table of API
calls that can contain URL arguments, so it is easy to add additional
API calls that we want to have URLs extracted from.  Add a couple of
others I've come across in testing.

BUG=253368, 275701

Review URL: https://chromiumcodereview.appspot.com/23545012