Refactor the daemon controller so that the callbacks are called on the caller thread.
This CL switches DaemonController to the same model that is used by PairingRegistry. DaemonController becomes a ref-counted object responsible for dispatching calls between threads. Platform-specific logic is implemented in non thread-safe delegate class. Most of the delegate's methods are synchronous and are running on a backgorund thread.

BUG=173509

Review URL: https://chromiumcodereview.appspot.com/23606019