[Files.app] Use generateTopLevelEntries() method to retrieve the list on the top level
This patch fixes the bug that the directory list is not drawn correctly sometimes.

Previously, DirectoryTree.updateSubDirectories() tries to retrieve an entries from the file system. But the directory structure of the top level is different from the file system, instead, we need to use the special list. This patch change the updateSubDirectories() to use the special list.

BUG=286421
TEST=manual
R=hirono@chromium.org

Review URL: https://codereview.chromium.org/23636014