sync: Implement per-type GetCommitIds
Convert the SyncerCommand GetCommitIdsCommand into a function named
GetCommitIds that serves the same purpose.  There was no benefit in
keeping this functionality in a SyncerCommand, and doing so would have
made the implementation of a per-type GetCommitIds more difficult.

Expose a function to get the IDs to commit for a single data type.  This
function is used in the implementation of GetCommitIdsCommand, but
nowhere else.  It will become more useful when we try to implement the
model type abstraction layer.

BUG=278484

Review URL: https://chromiumcodereview.appspot.com/23809005