PNaCl Coordinator: Run StreamEnd RPC even if StreamChunk fails
If the StreamChunk RPC fails, the coordinator currently reports an error
and aborts immedately, but this leaves no useful information for the
user or developer. Instead, call the StreamEnd RPC which returns a
string describing the error (but only if the RPC error was
reported by the application on the other end, and not an
error with the RPC itself).
Also disable the PnaclErrorHandling test until the translator change in
https://codereview.chromium.org/23753003/ rolls into Chrome
(since it, combined with this change, will improve the error message
reported by the translator).

R=jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3519
TEST=NaClBrowserTestPnacl.ErrorHandling

Review URL: https://chromiumcodereview.appspot.com/23946004