Telemetry: fix permission check when setting Android command line.
This is to fix a bug introduced by r221426 on the WebView telemetry bot.
The issue was related to the fact that ls -l can return no output if the
target folder exists but is empty.
This change makes the IsProtectedFile logic inside _SetCommandLineFile
more reliable, introducing the IsFileWritableOnDevice method in adb
commands.

BUG=284468
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/23477053