Fix white flashing on resize: fix wndclass brush, clear once to black
Fix setting of hbrBackground in WNDCLASS (which was a comedy of errors).
Because we're presenting with D3D (not GDI) it seems to be necessary to
clear the area that windows draws to during the NCPAINT, otherwise it
we get white artifacting. Black is treated as transparent by DWM code,
so by clearing, it correctly draws the window buttons.

R=sky@chromium.org
BUG=177115,266716

Review URL: https://chromiumcodereview.appspot.com/24088008