DumpWithoutCrashing() is looking up wrong breakpad_win entry point.
DumpWithoutCrashing() was looking up DumpProcess instead of
DumpProcessWithoutCrash. The former generates a dialog to restart
chrome while the latter does not. Neither actually crashes the
browser process.

Prior to the CL, anything invoking DumpWithoutCrashing() would appear
to the user as if a crash had occurred, allowing the user to start
a new browser process/window, but leaving the original up and running
(unless the caller subsequently crashed the browser itself).

R= rtenneti@chromium.org, mseaborn@chromium.org

Review URL: https://chromiumcodereview.appspot.com/23484022