Http Cache: Consider the case of calling StopCaching with the use of auth.
StopCaching may transition the transaction to passtrhough without taking full
account of the current state. As a result, if the transaction requires any
form of authentication, we may end up with reading_ being true and no response
headers. Now we account for that at the transaction destructor.

BUG=277011
TEST=net_unittests

R=gavinp@chromium.org

Review URL: https://chromiumcodereview.appspot.com/23726023