Only use the NativeViewHost::InstallClip functionality on Windows.
The installClip functionality seems to be an optimization for the renderer to only render content within a clipped area. This is implemented in Windows, but not on Aura in general. As such logging statements get produced for calling into a non existent feature. Removed the calls into that function since it is "optional" anyways.

BUG=288815
TEST=none

Review URL: https://chromiumcodereview.appspot.com/23883008