Remove the problematic call to GLSurface::GetBackingFrameBufferObject.
GLContextVirtual::IsCurrent doesn't need to call
GetBackingFrameBufferObject on a surface, the IsOffScreen check should
be enough.

The direct reason for this change is that GLContextVirtual::IsCurrent's
use of GetBackingFrameBufferObject interferes with the implementation of
that function in TextureImageTransportSurface.
TextureImageTransportSurface::GetBackingFrameBufferObject creates an FBO
if needed. Unfortunately GLContextVirtual::IsCurrent is called early in
GLContextVirtual's initialization, too early for
TextureImageTransportSurface::GetBackingFrameBufferObject to create the
frame buffer object because it needs a fully initialized context to
succeed.

BUG=

Review URL: https://chromiumcodereview.appspot.com/23480048