sync: Remove OverrideClientFieldsAfterCommit
I believe that this override takes effect only if the SessionTagRewriter
is invoked on the server side, and that applies only to Chrome on
Android clients with version <= M23.

Since we removed the server's "last say" on the positioning of items,
this function was the only thing blocking us from making
ProcessCommitResponseCommand non-model-changing.  Removing this function
clears the way for further refactorings that will reduce unnecessary
inter-thread communication.

BUG=284672

Review URL: https://chromiumcodereview.appspot.com/24019002