cc: Invalidate appropriate FBO attachment for offscreen rendering
The glInvalidateFramebuffer() uses different attachment names for the
default framebuffer and offscreen framebuffers. In the first case
we should use GL_COLOR_EXT, whereas offscreen framebuffers expect
GL_COLOR_ATTACHMENTn. This patch adds code to keep track of whether
GLRenderer is drawing to the default framebuffer or an offscreen one.
This information is used to pass the right attachment name when
clearing and invalidating the framebuffer.

BUG=288635

Review URL: https://chromiumcodereview.appspot.com/24089004