Simplified 'new connection' UI and disable for apps v1
An alternative mechanism for this was added in https://chromiumcodereview.appspot.com/23609010 but was considered hard to discover and, due to the way that <select> nodes are implemented, hiding it for apps v1 was tricky. This is a simpler UI that just adds a button next to the host name. Together with the app icon context menu option added in https://codereview.chromium.org/23536033/, this is good enough for feature parity with apps v1, although it's unlikely to be the final version.

This CL also adds transparency to the Reload icon, which previously had a white background.

BUG=252881
R=lambroslambrou@chromium.org

Review URL: https://codereview.chromium.org/23876010