IndexedDB: Keep backing store alive for time after connection closes
It appears to be common to close and re-open Indexed DB connections;
this can take hundreds of milliseconds even on fast hardware. Give
the backing store a short grace period (2 seconds) before closing.

This is not done in the case of a forced connection close if
that connection was the last one holding the backing store open, since
a forced-close is usually in response to wanting exclusive access to
the backing store.

BUG=153012

Review URL: https://chromiumcodereview.appspot.com/22793019