cc: Poll for state that triggers draws (synchronous compositor only)
This adds Scheduler::PollForAnticipatedDrawTriggers which is used
by the synchronous compositor to avoid requesting BeginImplFrames
when we might not actually draw new content, but still need to
monitor state that might trigger a draw at vsync intervals.

This can occur, for example, when the last swap used incomplete tiles
(checkerboards or low-quality) or when a commit is aborted and we
want to wait a vsync before starting the next commit.

BUG=284810

Review URL: https://chromiumcodereview.appspot.com/23463014