[iOS] History: Hide Clear Browsing Data button when keyboard is visible.
Existing code that handles this on Android and iOS 6 does not work on
iOS 7, so add some iOS-specific code that works on iOS 7.

Also deleted some obsolete iOS-specific code that was incorrect now
that history search is enabled.

BUG=271580
TEST=Open chrome://history. Ensure that the Clear Browsing Data button
is visible in both portrait and landscape when the keyboard is not
visible, and ensure the button is hidden when the keyboard *is* visible.

R=sergiu@chromium.org

Review URL: https://codereview.chromium.org/23558012