Adding audio unit tests for Android.
This CL adds support for a new set of audio unit tests intended for Android. Some of them are suitable for trybots but some are mainly intended for offline benchmarks/testing on different devices. The idea is that we shall use these tests to check if a certain device has potential of providing low-latency, high-quality audio in full duplex or not.

BUG=none
TEST=media_unittests

Review URL: https://chromiumcodereview.appspot.com/23296008