Implement window and view snapshots on Android
This patch implements Ui::GrabViewSnapshot and Ui::GrabWindowSnapshot on
Android. They both return a screen capture of the requested content
region. As on other platforms the URL bar will not be included in the
result, but note that any UI elements overlapping with the content
(e.g., the toolbar drop shadow) will be shown.
    
BUG=242299
TEST=tools/telemetry/run_tests --browser=android-chrome GpuTabTest -v -v -v

Review URL: https://chromiumcodereview.appspot.com/23632006